Fixed a bunch of DB stuff.

+++ RiotNet/Models/Block.cs
@@ -33,7 +33,7 @@ namespace RiotNet.Models
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         [JsonIgnore]
-        public int DatabaseId { get; set; }
+        public long DatabaseId { get; set; }
 #endif
     }
 }
