+++ RiotNet/Models/Player.cs
@@ -32,7 +32,7 @@ namespace RiotNet.Models
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         [JsonIgnore]
-        public int DatabaseId { get; set; }
+        public long DatabaseId { get; set; }
 #endif
     }
 }
